+++ apps/web/src/pages/api/auth/[...nextauth].ts
@@ -1,11 +1,12 @@
 /* eslint-disable no-param-reassign */
 import { NextApiHandler } from 'next';
+import { PrismaClient } from '@umamin/db';
 import NextAuth, { NextAuthOptions } from 'next-auth';
 import GoogleProvider from 'next-auth/providers/google';
 import DiscordProvider from 'next-auth/providers/discord';
 import { PrismaAdapter } from '@next-auth/prisma-adapter';
 
-import { prisma } from '@/db';
+const prisma = new PrismaClient();
 
 const options: NextAuthOptions = {
   debug: process.env.NODE_ENV === 'development',
